+++ fabric/src/main/java/io/github/flemmli97/improvedmobs/fabric/ImprovedMobsFabric.java
@@ -3,6 +3,7 @@ package io.github.flemmli97.improvedmobs.fabric;
 import io.github.flemmli97.improvedmobs.ImprovedMobs;
 import io.github.flemmli97.improvedmobs.config.Config;
 import io.github.flemmli97.improvedmobs.difficulty.DifficultyData;
+import io.github.flemmli97.improvedmobs.difficulty.IPlayerDifficulty;
 import io.github.flemmli97.improvedmobs.events.EventCalls;
 import io.github.flemmli97.improvedmobs.fabric.config.ConfigSpecs;
 import io.github.flemmli97.improvedmobs.fabric.events.EventHandler;
@@ -47,7 +48,7 @@ public class ImprovedMobsFabric implements ModInitializer {
         if (!ServerPlayNetworking.canSend(player, difficultyPacket))
             return;
         FriendlyByteBuf buf = PacketByteBufs.create();
-        buf.writeFloat(Config.CommonConfig.difficultyType == Config.DifficultyType.GLOBAL ? data.getDifficulty() : CrossPlatformStuff.INSTANCE.getPlayerDifficultyData(player).getDifficultyLevel());
+        buf.writeFloat(Config.CommonConfig.difficultyType == Config.DifficultyType.GLOBAL ? data.getDifficulty() : CrossPlatformStuff.INSTANCE.getPlayerDifficultyData(player).map(IPlayerDifficulty::getDifficultyLevel).orElse(0f));
         ServerPlayNetworking.send(player, difficultyPacket, buf);
     }
 
@@ -59,7 +60,7 @@ public class ImprovedMobsFabric implements ModInitializer {
         }
         PlayerLookup.all(server).forEach(player -> {
             if (!global)
-                buf.writeFloat(CrossPlatformStuff.INSTANCE.getPlayerDifficultyData(player).getDifficultyLevel());
+                buf.writeFloat(CrossPlatformStuff.INSTANCE.getPlayerDifficultyData(player).map(IPlayerDifficulty::getDifficultyLevel).orElse(0f));
             if (ServerPlayNetworking.canSend(player, difficultyPacket))
                 ServerPlayNetworking.send(player, difficultyPacket, buf);
         });